refactor(www): rename Rest to HTTP-API
