+++ config/config_test.go
@@ -46,12 +46,14 @@ func TestDefaultConfig(t *testing.T) {
 	assert.Zero(t, conf.Network.DefaultPort)
 
 	assert.False(t, conf.GRPC.Enable)
-	assert.False(t, conf.Rest.Enable)
+	assert.False(t, conf.HTML.Enable)
 	assert.False(t, conf.HTTP.Enable)
+	assert.False(t, conf.JSONRPC.Enable)
 
 	assert.Zero(t, conf.GRPC.Listen)
-	assert.Zero(t, conf.Rest.Listen)
+	assert.Zero(t, conf.HTML.Listen)
 	assert.Zero(t, conf.HTTP.Listen)
+	assert.Zero(t, conf.JSONRPC.Listen)
 }
 
 func TestMainnetConfig(t *testing.T) {
@@ -63,12 +65,14 @@ func TestMainnetConfig(t *testing.T) {
 	assert.Equal(t, 21888, conf.Network.DefaultPort)
 
 	assert.True(t, conf.GRPC.Enable)
-	assert.False(t, conf.Rest.Enable)
+	assert.False(t, conf.HTML.Enable)
 	assert.False(t, conf.HTTP.Enable)
+	assert.False(t, conf.JSONRPC.Enable)
 
 	assert.Equal(t, "127.0.0.1:50051", conf.GRPC.Listen)
-	assert.Equal(t, "127.0.0.1:8080", conf.Rest.Listen)
-	assert.Equal(t, "127.0.0.1:80", conf.HTTP.Listen)
+	assert.Equal(t, "127.0.0.1:80", conf.HTML.Listen)
+	assert.Equal(t, "127.0.0.1:8080", conf.HTTP.Listen)
+	assert.Equal(t, "127.0.0.1:8545", conf.JSONRPC.Listen)
 }
 
 func TestTestnetConfig(t *testing.T) {
@@ -80,12 +84,14 @@ func TestTestnetConfig(t *testing.T) {
 	assert.Equal(t, 21777, conf.Network.DefaultPort)
 
 	assert.True(t, conf.GRPC.Enable)
-	assert.True(t, conf.Rest.Enable)
-	assert.False(t, conf.HTTP.Enable)
+	assert.False(t, conf.HTML.Enable)
+	assert.True(t, conf.HTTP.Enable)
+	assert.True(t, conf.JSONRPC.Enable)
 
 	assert.Equal(t, "[::]:50052", conf.GRPC.Listen)
-	assert.Equal(t, "[::]:8080", conf.Rest.Listen)
-	assert.Equal(t, "[::]:80", conf.HTTP.Listen)
+	assert.Equal(t, "[::]:80", conf.HTML.Listen)
+	assert.Equal(t, "[::]:8080", conf.HTTP.Listen)
+	assert.Equal(t, "[::]:8545", conf.JSONRPC.Listen)
 }
 
 func TestLocalnetConfig(t *testing.T) {
@@ -97,12 +103,14 @@ func TestLocalnetConfig(t *testing.T) {
 	assert.Equal(t, 0, conf.Network.DefaultPort)
 
 	assert.True(t, conf.GRPC.Enable)
-	assert.True(t, conf.Rest.Enable)
+	assert.True(t, conf.HTML.Enable)
 	assert.True(t, conf.HTTP.Enable)
+	assert.True(t, conf.JSONRPC.Enable)
 
 	assert.Equal(t, "[::]:50052", conf.GRPC.Listen)
-	assert.Equal(t, "[::]:8080", conf.Rest.Listen)
-	assert.Equal(t, "[::]:0", conf.HTTP.Listen)
+	assert.Equal(t, "[::]:0", conf.HTML.Listen)
+	assert.Equal(t, "[::]:8080", conf.HTTP.Listen)
+	assert.Equal(t, "[::]:8545", conf.JSONRPC.Listen)
 }
 
 func TestLoadFromFile(t *testing.T) {
