+++ util/logger/config.go
@@ -26,10 +26,9 @@ func DefaultConfig() *Config {
 	conf.Levels["_sync"] = "error"
 	conf.Levels["_pool"] = "error"
 	conf.Levels["_http"] = "info"
+	conf.Levels["_html"] = "info"
 	conf.Levels["_grpc"] = "info"
-	conf.Levels["_nonomsg"] = "info"
 	conf.Levels["_jsonrpc"] = "info"
-	conf.Levels["_rest"] = "info"
 	conf.Levels["_zmq"] = "info"
 	conf.Levels["_firewall"] = "warn"
