+++ www/grpc/buf/openapi.config.yaml
@@ -4,7 +4,7 @@ openapiOptions:
   file:
     - file: "transaction.proto"
       option:
-        basePath: "/rest/api"
+        basePath: "/http/api"
         info:
           title: Pactus APIs
           version: "2.0"
