+++ www/html/config.go
@@ -1,0 +1,20 @@
+package html
+
+type Config struct {
+	Enable      bool   `toml:"enable"`
+	Listen      string `toml:"listen"`
+	EnablePprof bool   `toml:"enable_pprof"`
+}
+
+func DefaultConfig() *Config {
+	return &Config{
+		Enable:      false,
+		Listen:      "",
+		EnablePprof: false,
+	}
+}
+
+// BasicCheck performs basic checks on the configuration.
+func (*Config) BasicCheck() error {
+	return nil
+}
