+++ www/html/html_test.go
@@ -1,4 +1,4 @@
-package http
+package html
 
 import (
 	"context"
