+++ www/http/swagger-ui/swagger-ui-standalone-preset.js
